Add files via upload
The practice of assisting people in locating and pursuing occupations that are a good fit for their abilities, interests, and goals is known as career counselling. Although it can be a useful tool for individuals of all ages, students and young adults just starting their professions should pay particular attention to it.

Career coaching is one of the many professions that artificial intelligence (AI) is quickly transforming. Compared to traditional career counselling services, AI-powered platforms can offer students and young people more timely, tailored guidance.